feat: swell tooltip in FarmV3Card, move SwellTooltip to global
