+++ apps/web/src/views/Farms/components/FarmCard/CardHeading.tsx
@@ -16,7 +16,9 @@ import {
 import { FeeAmount } from '@pancakeswap/v3-sdk'
 import { FarmWidget } from '@pancakeswap/widgets-internal'
 import { GiftTooltip } from 'components/GiftTooltip/GiftTooltip'
+import { SwellTooltip } from 'components/SwellTooltip/SwellTooltip'
 import { TokenPairImage } from 'components/TokenImage'
+import { useHasSwellReward } from 'hooks/useHasSwellReward'
 import { styled } from 'styled-components'
 import { Address, isAddressEqual } from 'viem'
 import { bsc } from 'viem/chains'
@@ -79,6 +81,7 @@ const CardHeading: React.FC<React.PropsWithChildren<ExpandableSectionProps>> = (
   const { t } = useTranslation()
   const chainId = useChainId()
   const isReady = multiplier !== undefined || bCakeWrapperAddress
+  const hasSwellReward = useHasSwellReward(lpAddress ?? '0x0', 'FARM')
   const multiplierTooltipContent = FarmMultiplierInfo({
     farmCakePerSecond: farmCakePerSecond ?? '-',
     totalMultipliers: totalMultipliers ?? '-',
@@ -110,6 +113,7 @@ const CardHeading: React.FC<React.PropsWithChildren<ExpandableSectionProps>> = (
                 />
               </Box>
             ) : null}
+
             {chainId === bsc.id &&
             lpAddress &&
             isAddressEqual(lpAddress, '0xdD82975ab85E745c84e497FD75ba409Ec02d4739') ? (
@@ -142,6 +146,12 @@ const CardHeading: React.FC<React.PropsWithChildren<ExpandableSectionProps>> = (
           <Skeleton mb="4px" width={60} height={18} />
         )}
         <AutoRow gap="4px" justifyContent="flex-end">
+          {hasSwellReward && (
+            <>
+              &nbsp;
+              <SwellTooltip />
+            </>
+          )}
           {isReady && isStable ? <StableFarmTag /> : version === 2 ? <V2Tag /> : null}
           {isReady && version === 3 && <V3FeeTag feeAmount={feeAmount} />}
           {isReady && isCommunityFarm && <FarmAuctionTag mr="-4px" />}
