+++ apps/web/src/views/PositionManagers/components/DuoTokenVaultCard.tsx
@@ -4,6 +4,7 @@ import { Card, CardBody } from '@pancakeswap/uikit'
 import { getBalanceAmount } from '@pancakeswap/utils/formatBalance'
 import { FeeAmount } from '@pancakeswap/v3-sdk'
 import BigNumber from 'bignumber.js'
+import { SwellTooltip } from 'components/SwellTooltip/SwellTooltip'
 import { useHasSwellReward } from 'hooks/useHasSwellReward'
 import { PropsWithChildren, ReactNode, memo, useMemo } from 'react'
 import { styled } from 'styled-components'
@@ -16,7 +17,6 @@ import { CardTitle } from './CardTitle'
 import { ExpandableSection } from './ExpandableSection'
 import { LiquidityManagement } from './LiquidityManagement'
 import { ManagerInfo } from './ManagerInfo'
-import { SwellTooltip } from './SwellTooltip'
 import { VaultInfo } from './VaultInfo'
 import { VaultLinks } from './VaultLinks'
 import { YieldInfo } from './YieldInfo'
