+++ apps/web/src/views/PositionManagers/components/TableView/index.tsx
@@ -4,6 +4,7 @@ import { CurrencyAmount } from '@pancakeswap/sdk'
 import { getBalanceAmount } from '@pancakeswap/utils/formatBalance'
 import { useQuery } from '@tanstack/react-query'
 import BigNumber from 'bignumber.js'
+import { SwellTooltip } from 'components/SwellTooltip/SwellTooltip'
 import { usePositionManagerAdepterContract } from 'hooks/useContract'
 import { useHasSwellReward } from 'hooks/useHasSwellReward'
 import { useBCakeBoostLimitAndLockInfo } from 'views/Farms/components/YieldBooster/hooks/bCakeV3/useBCakeV3Info'
@@ -38,7 +39,6 @@ import { FarmCell } from './FarmCell'
 import { AprMobileCell, CellInner, FarmMobileCell, StyledTr } from './Styled'
 
 import { TIME_WINDOW_DEFAULT, TIME_WINDOW_FALLBACK } from '../../hooks/useFetchApr'
-import { SwellTooltip } from '../SwellTooltip'
 
 interface Props {
   config: PCSDuoTokenVaultConfig
